Add Scene / Group commands in functions.sh
I've noticed that the plugin does not allow us to send on/off commands for scenes or groups favorited in Domoticz. I looked what was going wrong, and it seems that the curl call for Domoticz API is sending switchlight in pg_dz_switch function. Though, groups or scenes activation results in an error, since it seems to need switchscene command to be passed according to Domoticz API.

I'm a poor programmer, and my modifications are awful for sure, but i've made a few changes to functions.sh in order to make it work properly regarding scenes and groups. If someone wants to beta-test it and improve it, you're really welcome !